+++ main.lua
@@ -1,202 +1,13 @@
 -- Tower Defence v1.14.10
 -- By Jeren (Code), Jasper (Art & Sound), Jimmy (Art & Music)
+
+-- Game Data and Reset Data
 #include src/data.lua
 -- Classes
-Enemy = {}
-function Enemy:new(location, enemy_data)
-  obj = { 
-    x = location[1], y = location[2],
-    hp = enemy_data.hp, 
-    step_delay = enemy_data.step_delay,
-    current_step = 0,
-    is_frozen = false,
-    frozen_tick = 0,
-    burning_tick = 0,
-    gfx = enemy_data.sprite_index,
-    reward = enemy_data.reward,
-    damage = enemy_data.damage,
-    pos = 1
-  }
-  setmetatable(obj, self)
-  self.__index = self
-  return obj
-end
-function Enemy:step()
-  self.current_step = (self.current_step + 1) % self.step_delay
-  if (self.current_step ~= 0) return false
-
-  if self.burning_tick > 0 then 
-    self.burning_tick -= 1
-    self.hp -= 2
-    local px, py, _ = Enemy.get_pixel_location(self)
-    add(particles, Particle:new(px, py, true, Animator:new(particle_data.burn, false)))
-  end
-
-  if (not self.is_frozen) return true 
-  self.frozen_tick = max(self.frozen_tick - 1, 0)
-  if (self.frozen_tick ~= 0) return false
-  self.is_frozen = false
-  return true
-end
-function Enemy:get_pixel_location()
-  local n, prev = pathing[self.pos], {x=map_data[loaded_map].enemy_spawn_location[1], y=map_data[loaded_map].enemy_spawn_location[2]}
-  if (self.pos - 1 >= 1) prev = pathing[self.pos-1]
-  local px, py = self.x * 8, self.y * 8
-  if not self.is_frozen then 
-    px = lerp(prev.x * 8, n.x * 8, self.current_step / self.step_delay)
-    py = lerp(prev.y * 8, n.y * 8, self.current_step / self.step_delay)
-  end
-  return px, py, n
-end
-function Enemy:draw()
-  if (self.hp <= 0) return
-  local px, py, n = Enemy.get_pixel_location(self)
-  local dir = {normalize(n.x - self.x), normalize(n.y - self.y)}
-  spr(parse_direction(self.gfx, dir), px, py, 1, 1, get_flip_direction(dir))
-end
-
-Tower = {}
-function Tower:new(dx, dy, tower_template_data, direction)
-  obj = { 
-    x = dx, y = dy,
-    dmg = tower_template_data.damage,
-    radius = tower_template_data.radius, 
-    attack_delay = tower_template_data.attack_delay,
-    current_attack_ticks = 0,
-    cost = tower_template_data.cost,
-    type = tower_template_data.type,
-    dir = direction,
-    single_hit = tower_template_data.single_tile_hit_only,
-    animator = Animator:new({
-      sprite_data = tower_template_data.sprite_data,
-      ticks_per_frame = tower_template_data.ticks_per_frame
-    }, true),
-  }
-  add(animators, obj.animator)
-  setmetatable(obj, self)
-  self.__index = self 
-  return obj 
-end
-function Tower:attack()
-  self.current_attack_ticks = (self.current_attack_ticks + 1) % self.attack_delay
-  if (self.current_attack_ticks > 0) return
-
-  if self.type == "tack" then
-    Tower.apply_damage(self, Tower.nova_collision(self))
-  elseif self.type == "rail" then
-    Tower.apply_damage(self, Tower.raycast(self))
-  elseif self.type == "frontal" then 
-    Tower.freeze_enemies(self, Tower.frontal_collision(self))
-  elseif self.type == "floor" then 
-    local hits = {}
-    add_enemy_at_to_table(self.x, self.y, hits)
-    foreach(hits, function(enemy) enemy.burning_tick += self.dmg end)
-  end
-end
-function Tower:raycast()
-  if (self.dir[1] == 0 and self.dir[2] == 0) return nil
-  local hits = {}
-  for i=1, self.radius do 
-    add_enemy_at_to_table(self.x + i * self.dir[1] , self.y + i * self.dir[2], hits)
-  end
-  if (#hits > 0) raycast_spawn(self.x, self.y, self.radius, self.dir, particle_data.spark)
-  return hits
-end
-function Tower:nova_collision()
-  local hits = {}
-  for y=-self.radius, self.radius do
-    for x=-self.radius, self.radius do
-      if (x ~= 0 or y ~= 0) add_enemy_at_to_table(self.x + x, self.y + y, hits, self.single_hit)
-    end
-  end
-  if (#hits > 0) nova_spawn(self.x, self.y, self.radius, particle_data.blade)
-  return hits
-end
-function Tower:frontal_collision()
-  local hits = {}
-  local fx, fy, flx, fly, ix, iy = parse_frontal_bounds(self.dir[1], self.dir[2], self.radius)
-  for y=fy, fly, iy do
-    for x=fx, flx, ix do
-      if (x ~= 0 or y ~= 0) add_enemy_at_to_table(self.x + x, self.y + y, hits)
-    end
-  end
-  if (#hits > 0) frontal_spawn(self.x, self.y, self.radius, self.dir, particle_data.frost)
-  return hits
-end
-function Tower:apply_damage(targets)
-  for _, enemy in pairs(targets) do
-    if (enemy.hp > 0) enemy.hp -= self.dmg
-  end
-end
-function Tower:freeze_enemies(targets)
-  for _, enemy in pairs(targets) do
-    if not enemy.is_frozen then 
-      enemy.is_frozen = true
-      enemy.frozen_tick = self.dmg
-    end 
-  end
-end
-function Tower:draw()
-  local id = parse_direction(Animator.sprite_id(self.animator), self.dir)
-  spr(id, self.x * 8, self.y * 8, 1, 1, get_flip_direction(self.dir))
-end
-
-Particle = {}
-function Particle:new(dx, dy, pixel_perfect, animator_)
-  obj = {
-    x = dx, y = dy,
-    is_pxl_perfect = pixel_perfect or false,
-    animator = animator_,
-  }
-  setmetatable(obj, self)
-  self.__index = self
-  return obj
-end
-function Particle:tick()
-  return Animator.update(self.animator)
-end
-function Particle:draw()
-  if (Animator.finished(self.animator)) return 
-  local id = Animator.sprite_id(self.animator)
-  if self.is_pxl_perfect then 
-    spr(id, self.x, self.y)
-  else
-    spr(id, self.x*8, self.y*8)
-  end
-end
-
-Animator = {}
-function Animator:new(data, continuous_)
-  obj = {
-    sprite_data = data.sprite_data,
-    animation_frame = 1,
-    frame_duration = data.ticks_per_frame,
-    tick = 0,
-    continuous = continuous_
-  }
-  setmetatable(obj, self)
-  self.__index = self
-  return obj
-end
-function Animator:update()
-  self.tick = (self.tick + 1) % self.frame_duration
-  if (self.tick ~= 0) return false
-  if Animator.finished(self) then 
-    if (self.continuous) Animator.reset(self)
-    return true
-  end
-  self.animation_frame += 1
-  return false
-end
-function Animator:finished()
-  return self.animation_frame >= #self.sprite_data
-end
-function Animator:sprite_id()
-  return self.sprite_data[self.animation_frame]
-end
-function Animator:reset()
-  self.animation_frame = 1
-end
+#include src/enemy.lua
+#include src/tower.lua
+#include src/particle.lua
+#include src/animator.lua
 
 -- Pico8
 function _init()
